+++ packages/audioplayers/lib/audioplayers.dart
@@ -7,9 +7,9 @@ export 'package:audioplayers_platform_interface/src/api/player_state.dart';
 export 'package:audioplayers_platform_interface/src/api/release_mode.dart';
 
 export 'src/audio_cache.dart';
+export 'src/audio_log_level.dart';
+export 'src/audio_logger.dart';
 export 'src/audio_pool.dart';
 export 'src/audioplayer.dart';
 export 'src/global_audio_scope.dart';
-export 'src/log_level.dart';
-export 'src/logger.dart';
 export 'src/source.dart';
