+++ packages/audioplayers/lib/src/global_audio_scope.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
 
-import 'package:audioplayers/src/log_level.dart';
-import 'package:audioplayers/src/logger.dart';
+import 'package:audioplayers/src/audio_log_level.dart';
+import 'package:audioplayers/src/audio_logger.dart';
 import 'package:audioplayers_platform_interface/audioplayers_platform_interface.dart';
 
 /// Handle global audio scope like calls and events concerning all AudioPlayers.
